Fixes bug and improves store list refresh upon sync.
When the logs from another device are updated on disk, the MOC needs to
flush the cache for that database. Now using a “blink” approach where
the URL of the store is set again, instead of a “swap” approach where
the persistent store is removed and a new one is created. There was
also a bug where the list of read-only databases (persistent stores)
would be continuously growing.