Triggers a "blink" of the read-write log database of the local device after saving to disk.
The “blink” is necessary for Dropbox to upload the modified database to
their servers. The “blinking” is done by simply setting again the store
URL to the same value as before. It is not needed on iOS where the
files are  under the control of the app and will need to be manually
uploaded to online file services.