Using the correct 64-bit signed integer type for timestamps.
The timestamps are stored in CoreData as signed 64-bit integer, no
matter what we use with the NSNumber content, which is cast anyway when
written to SQLite. The “distant past” and “distant future” timestamps
would however have a different value on 32-bit iOS. The distant future
would in fact correspond to Jan 1, 2001, around noon.