Changed the appendChanges: method
It is now an insertChanges: method. It can insert at any position, or it can do an append only insert. Importantly, in both cases, it now checks for duplicates, and will allow a new change with the same timestamp as the existing maximum timestamp (provided the new change is unique).