Added ability to disabled in-memory caching in PARStore.
You call disableInMemoryCache before loading. This sets the in-memory dictionary to nil, and sets a flag.

In general, all the same procedures are undertaken without the cache as with it, but the values are not actually stored in a dictionary. We do this so that key times and other aspects such as notifications work as expected.