Adding 2 more tests that reveal an issue with getting log rows in proper timestamp order in all cases.
It looks like the sort descriptors are pretty much ignored when a fetch
request is configured to return just the object IDs. This means the
rows are not returned in the expected reverse timestamp order. The
issue seems close to 100% reproducible in the 2 new tests just added.
It’s not entirely clear it’s 100% though, as it will likely depends on
the timing and the order in which things are stored in the various
database files.