Fixing issue with merge implementation that would not apply changes made during the merge in the proper order.
See test case as well. I checked that it fails before the fix, and
succeeds with the change. The issue was that the implementation of
load/sync was modifying the memory cache asynchronously, but the merge
implementation operates within the memory queue, and was assuming all
changes had been applied. To avoid blocking the database queue if not
needed, we could also check in the `_sync` implementation if we are
already within the memory queue, and only use a synchronous call if we
are, otherwise make the call asynchronous. The merge context is very
rarely encountered, so a synchronous call should remain the exception.
On the other hand, it’s only relevant for the initial “load” of the
data into the memory cache.